perfetto: move murmurhash to base and refactor base::Hash
* move murmurhash to base
* move base::hash FNV implementation -> base::FnvHash
* make base::Hash a shim for base::FnvHash
* add murmurhash unittest file
