+++ Android.bp
@@ -12310,12 +12310,13 @@ filegroup {
         "src/base/flags_unittest.cc",
         "src/base/flat_hash_map_unittest.cc",
         "src/base/flat_set_unittest.cc",
+        "src/base/fnv_hash_unittest.cc",
         "src/base/getopt_compat_unittest.cc",
-        "src/base/hash_unittest.cc",
         "src/base/intrusive_list_unittest.cc",
         "src/base/intrusive_tree_unittest.cc",
         "src/base/logging_unittest.cc",
         "src/base/metatrace_unittest.cc",
+        "src/base/murmur_hash_unittest.cc",
         "src/base/no_destructor_unittest.cc",
         "src/base/paged_memory_unittest.cc",
         "src/base/periodic_task_unittest.cc",
