+++ BUILD
@@ -836,11 +836,13 @@ perfetto_filegroup(
         "include/perfetto/ext/base/flags.h",
         "include/perfetto/ext/base/flags_list.h",
         "include/perfetto/ext/base/flat_hash_map.h",
+        "include/perfetto/ext/base/fnv_hash.h",
         "include/perfetto/ext/base/getopt.h",
         "include/perfetto/ext/base/getopt_compat.h",
         "include/perfetto/ext/base/hash.h",
         "include/perfetto/ext/base/metatrace.h",
         "include/perfetto/ext/base/metatrace_events.h",
+        "include/perfetto/ext/base/murmur_hash.h",
         "include/perfetto/ext/base/no_destructor.h",
         "include/perfetto/ext/base/paged_memory.h",
         "include/perfetto/ext/base/periodic_task.h",
