+++ include/perfetto/ext/base/BUILD.gn
@@ -29,11 +29,13 @@ source_set("base") {
     "flags.h",
     "flags_list.h",
     "flat_hash_map.h",
+    "fnv_hash.h",
     "getopt.h",
     "getopt_compat.h",
     "hash.h",
     "metatrace.h",
     "metatrace_events.h",
+    "murmur_hash.h",
     "no_destructor.h",
     "paged_memory.h",
     "periodic_task.h",
