+++ src/base/BUILD.gn
@@ -212,11 +212,12 @@ perfetto_unittest_source_set("unittests") {
     "flags_unittest.cc",
     "flat_hash_map_unittest.cc",
     "flat_set_unittest.cc",
+    "fnv_hash_unittest.cc",
     "getopt_compat_unittest.cc",
-    "hash_unittest.cc",
     "intrusive_list_unittest.cc",
     "intrusive_tree_unittest.cc",
     "logging_unittest.cc",
+    "murmur_hash_unittest.cc",
     "no_destructor_unittest.cc",
     "paged_memory_unittest.cc",
     "periodic_task_unittest.cc",
