+++ src/trace_processor/util/BUILD.gn
@@ -92,14 +92,6 @@ source_set("protozero_to_text") {
   ]
 }
 
-source_set("murmur_hash") {
-  sources = [ "murmur_hash.h" ]
-  deps = [
-    "../../../gn:default_deps",
-    "../../base",
-  ]
-}
-
 source_set("protozero_to_json") {
   sources = [
     "protozero_to_json.cc",
